Conflicts:
	app/MediaUploader/Sitecore.Mobile.MediaUploader.xcodeproj/project.pbxproj
	app/MediaUploader/Sitecore.Mobile.MediaUploader/Data/sc_ItemsForUploadManager.h
	app/MediaUploader/Sitecore.Mobile.MediaUploader/Data/sc_ItemsForUploadManager.mm
	app/MediaUploader/Sitecore.Mobile.MediaUploader/Sitecore.Mobile.MediaUploader-Prefix.pch
	app/MediaUploader/Sitecore.Mobile.MediaUploader/sc_QuickImageViewController.h
	app/MediaUploader/Sitecore.Mobile.MediaUploader/sc_SettingsViewController.mm
	app/MediaUploader/Sitecore.Mobile.MediaUploader/sc_Upload2ViewController.mm
	app/MediaUploader/Sitecore.Mobile.MediaUploader/sc_UploadViewController.m
	lib/MUUploadManager/MUUploadManager/MUImageHelper.mm

+++ app/MediaUploader/Sitecore.Mobile.MediaUploader/sc_ViewsHelper.h
@@ -9,6 +9,8 @@
 #import <Foundation/Foundation.h>
 
 @interface sc_ViewsHelper : NSObject
-+(UIViewController* )reloadParentController:(UINavigationController*) navigationController levels:(int) levels;
+
++(UIViewController*)reloadParentController:(UINavigationController*)navigationController
+                                    levels:(NSUInteger)levels;
 
 @end
